requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/LLM Integration
openai>=1.10.0
# ollama>=0.1.0  # Uncomment for local LLM support

# GitHub & Git Integration
PyGithub>=2.1.1
gitpython>=3.1.40

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# Task Scheduling
apscheduler>=3.10.4

# CLI & UI
typer>=0.9.0
rich>=13.7.0
click>=8.1.7

# Template Engine
jinja2>=3.1.3

# Data Processing & Analytics
pandas>=2.1.4
plotly>=5.18.0

# Notifications (Optional)
# sendgrid>=6.11.0
# twilio>=8.11.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.1
ruff>=0.1.9
mypy>=1.8.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
